Исправил ошибки

+++ server/database.py
@@ -1,17 +1,18 @@
 import psycopg2
-import datetime
 from lib import Task
 
 params = dict(dbname='default', user="evgenabramov", password="12345678", host="localhost")
 
 
-def connect_database(name):
+def connect_database(name, user, password):
     params['dbname'] = name
+    params['user'] = user
+    params['password'] = password
     with psycopg2.connect(**params) as conn:
         cur = conn.cursor()
         cur.execute('''
             CREATE TABLE IF NOT EXISTS tasks(
-                task_name VARCHAR(20),
+                task_name VARCHAR(40),
                 due_date DATE,
                 description VARCHAR(1000),
                 completed BOOLEAN DEFAULT False
@@ -23,19 +24,29 @@ def add_task(task):
     with psycopg2.connect(**params) as conn:
         cur = conn.cursor()
         cur.execute('INSERT INTO tasks (task_name, due_date, description) VALUES (%s, %s, %s);',
-                    (task.name, datetime.datetime.strftime(task.due_date, '%Y-%m-%d'), task.description))
+                    (task.name, task.get_database_representation(), task.description))
 
 
 def delete_task(name):
     with psycopg2.connect(**params) as conn:
         cur = conn.cursor()
+        cur.execute('SELECT * FROM tasks WHERE task_name = %s;', [name])
+        rows = cur.fetchall()
+        if len(rows) == 0:
+            return False
         cur.execute('DELETE FROM tasks WHERE task_name = %s;', [name])
+        return True
 
 
 def mark_completed(name):
     with psycopg2.connect(**params) as conn:
         cur = conn.cursor()
+        cur.execute('SELECT * FROM tasks WHERE task_name = %s;', [name])
+        rows = cur.fetchall()
+        if len(rows) == 0:
+            return False
         cur.execute('UPDATE tasks SET completed = True WHERE task_name = %s;', [name])
+        return True
 
 
 def show_tasks(latest_date, with_completed):
